Revert deprecations, make DList abstract
After thinking about it, I realized it wasn't necessary to deprecate these
functions. Some people prefer to use the non-type-class versions, so I shouldn't
force them otherwise. It's best to be more conservative in the changes, too.

It's useful to compare this commit with 349293ef171b7aa151b2475acd773f50a2c67169
to see the actual changes up to this point.